+++ lib/main.dart
@@ -1,25 +1,39 @@
-import 'package:assignment/presentation/screens/home_screen.dart';
-import 'package:assignment/presentation/screens/splash_screen.dart';
+import 'package:assignment/logic/cubit/app_theme_cubit.dart';
+import 'package:assignment/presentation/router/app_router.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sizer/sizer.dart';
 
 import 'core/constants/strings.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+  MyApp({super.key});
+  bool setColor = false;
 
   @override
   Widget build(BuildContext context) {
-    return Sizer(builder: (context, constraints, orientation){
-      return const MaterialApp(
-        title: Strings.appTitle,
-        debugShowCheckedModeBanner: false,
-        home: SplashScreen(),
-      );
-    });
+    return MultiBlocProvider(
+      providers: [
+        BlocProvider(
+          create: (context) => AppThemeCubit(),
+        ),
+      ],
+      child: Sizer(builder: (context, constraints, orientation) {
+        if(!setColor){
+          BlocProvider.of<AppThemeCubit>(context).setLightTheme();
+          setColor = true;
+        }
+        return const MaterialApp(
+          title: Strings.appTitle,
+          debugShowCheckedModeBanner: false,
+          initialRoute: AppRouter.splash,
+          onGenerateRoute: AppRouter.onGenerateRoute,
+        );
+      }),
+    );
   }
 }