+++ lib/presentation/utils/custom_text.dart
@@ -1,4 +1,6 @@
+import 'package:assignment/logic/cubit/app_theme_cubit.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:sizer/sizer.dart';
 
@@ -19,6 +21,7 @@ class CustomText extends StatefulWidget {
   final Color? textColor;
   final TextAlign? textAlign;
   final FontWeight? fontWeight;
+
   @override
   State<CustomText> createState() => _CustomTextState();
 }
@@ -26,14 +29,18 @@ class CustomText extends StatefulWidget {
 class _CustomTextState extends State<CustomText> {
   @override
   Widget build(BuildContext context) {
-    return Text(
-      widget.textString,
-      textAlign: widget.textAlign??TextAlign.start,
-      style: GoogleFonts.poppins(
-          color: widget.textColor,
-          fontSize: widget.textFontSize??12.0.sp,
-          fontWeight: widget.fontWeight??FontWeight.normal,
-      ),
+    return BlocBuilder<AppThemeCubit, AppThemeState>(
+      builder: (context, appThemeState) {
+        return Text(
+          widget.textString,
+          textAlign: widget.textAlign ?? TextAlign.start,
+          style: GoogleFonts.poppins(
+            color: widget.textColor??(appThemeState as AppThemeSet).themeClass.textColor_1,
+            fontSize: widget.textFontSize ?? 12.0.sp,
+            fontWeight: widget.fontWeight ?? FontWeight.normal,
+          ),
+        );
+      },
     );
   }
 }